Renamed fdbbench to flowbench
