+++ flowbench/BenchIterate.cpp
@@ -20,12 +20,12 @@
 
 #include "benchmark/benchmark.h"
 
-#include "fdbbench/GlobalData.h"
 #include "fdbclient/CommitTransaction.h"
 #include "fdbclient/FDBTypes.h"
 #include "fdbclient/MutationList.h"
 #include "flow/Arena.h"
 #include "flow/FastAlloc.h"
+#include "flowbench/GlobalData.h"
 
 void populate(Standalone<VectorRef<MutationRef>>& mutations, size_t items, size_t size, KeyRef key, ValueRef value) {
 	mutations = Standalone<VectorRef<MutationRef>>{};
