+++ flowbench/BenchPopulate.cpp
@@ -20,11 +20,11 @@
 
 #include "benchmark/benchmark.h"
 
-#include "fdbbench/GlobalData.h"
 #include "fdbclient/CommitTransaction.h"
 #include "fdbclient/FDBTypes.h"
 #include "flow/Arena.h"
 #include "flow/FastAlloc.h"
+#include "flowbench/GlobalData.h"
 
 static constexpr bool EMPLACE_BACK = true;
 static constexpr bool PUSH_BACK = false;
