+++ flowbench/GlobalData.cpp
