+++ flowbench/flowbench.actor.cpp
@@ -1,5 +1,5 @@
 /*
- * fdbbench.actor.cpp
+ * flowbench.actor.cpp
  *
  * This source file is part of the FoundationDB open source project
  *
